size_t -> int

+++ audioFeature/src/stft.c
@@ -6,17 +6,17 @@
 
 #include "stft.h"
 
-#include <stdio.h>
+#include <stdlib.h>
 #include <math.h>
 #include <gsl/gsl_fft_complex.h>
 
-double hann_window(size_t n, size_t size) {
+double hann_window(int n, int size) {
   return 0.5 * (1 - cos(2 * M_PI * n / (size - 1)));
 }
 
-double (*window)(size_t, size_t) = hann_window;
+double (*window)(int, int) = hann_window;
 
-TimeFreq* alloc_tf(size_t window_size, size_t nos) {
+TimeFreq* alloc_tf(int window_size, int nos) {
   TimeFreq* tf = malloc(sizeof(TimeFreq));
   if (tf == NULL) {
     return NULL;
@@ -28,11 +28,11 @@ TimeFreq* alloc_tf(size_t window_size, size_t nos) {
     return NULL;
   }
 
-  size_t i;
+  int i;
   for (i = 0; i < nos; i++) {
     data[i] = malloc(2 * window_size * sizeof(double));
     if (data[i] == NULL) {
-      size_t j;
+      int j;
       for (j = 0; j < i; j++) {
         free(data[j]);
       }
@@ -50,7 +50,7 @@ TimeFreq* alloc_tf(size_t window_size, size_t nos) {
 }
 
 void free_tf(TimeFreq* tf) {
-  size_t i;
+  int i;
   for (i = 0; i < tf->nos; i++) {
     free(tf->data[i]);
   }
@@ -75,12 +75,12 @@ void set_value(Spectra spec, int index, double real, double imag) {
   spec[index * 2 + 1] = imag;
 }
 
-size_t number_of_spectrum(size_t length, size_t window_size, size_t shift) {
+int number_of_spectrum(int length, int window_size, int shift) {
   return (length - window_size) / shift + 1;
 }
 
-void stft(double* data, size_t length, size_t window_size,
-    size_t shift, TimeFreq* result) {
+void stft(double* data, int length, int window_size, int shift,
+    TimeFreq* result) {
   result->window_size = window_size;
   result->shift = shift;
   result->nos = number_of_spectrum(length, window_size, shift);
@@ -88,7 +88,7 @@ void stft(double* data, size_t length, size_t window_size,
   for (i = 0; i < result->nos; i++) {
     Spectra s = get_spectra(result, i);
     double* offset = data + i * shift;
-    size_t j;
+    int j;
     for (j = 0; j < window_size; j++) {
       set_value(s, j, offset[j] * window(j, window_size), 0);
     }
